code/trunk/event_horizon/requirements.txt
Django==1.5.1
PyXML==0.8.4
PyYAML==3.10
South==0.8.1
distribute==0.6.15
django-mssql==1.3.1
elementtree==1.2.7-20070827-preview
epydoc==3.0.1
fpconst==0.7.2
gdata==2.0.18
google-api-python-client==1.2
googleads==1.0.2
httplib2==0.8
isapi-wsgi==0.4.2
mock==1.0.1
oauth2client==1.2
oauthlib==0.6.1
pypm==1.3.4
python-dateutil==2.2
pythonselect==1.3
pytz==2013.8
pywin32==214
six==1.4.1
suds-jurko==0.6
tornado==3.1
tweepy==2.1
virtualenv==1.6.1
wsgiref==0.1.2
